% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{compare_nested}
\alias{compare_nested}
\title{Compares nested models using partial log-likelihood}
\usage{
compare_nested(crfull, crreduced)
}
\arguments{
\item{crfull}{Cox Regression object for the full model}

\item{crreduced}{Cox Regression object for the reduced model}
}
\value{
data frame with the statistic and p-value
}
\description{
Compares nested models using partial log-likelihood
}
