% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{cv}
\alias{cv}
\title{Performs an iteration of cross-validation}
\usage{
cv(variables, train, test)
}
\arguments{
\item{variables}{a string of the variables to use in the model ("new_user + text_length + ...")}

\item{train}{training set to build the model on}

\item{test}{test set to predict on}
}
\value{
data frame with prediction performance metrics for the training and test set
}
\description{
Fits a Cox proportional hazards model to the given training set and predicts on the input
test set. Performance metrics for predicted hazard ratios for training and test sets are calculated.
}
