% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_functions.R
\name{exploratory_setup}
\alias{exploratory_setup}
\title{Set up iFixit Answers data for exploratory analysis}
\usage{
exploratory_setup()
}
\value{
Returns Answers data.
}
\description{
The function was used to set up all variables for exploratory analysis including variables not
included in the final cox regression model (e.g. num_freq_tags, text_till_punct...). Takes no parameters.
}
