% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_failure_functions.R
\name{fit_model}
\alias{fit_model}
\title{Fit cox regression model}
\usage{
fit_model(data, summary = FALSE)
}
\arguments{
\item{data}{The data to use in fitting the model.}

\item{summary}{If true, this function will print a summary of the model (e.g. statistics, coefficients).
Default is set to false.}
}
\value{
Returns the cox regression model with 19 independent predictors. See the documentation on the
variable_setup function for details on what each predictor represents.
}
\description{
This function sets up variables in the input data using the variable_setup function,
and then fits and returns the model.
}
\details{
\itemize {
  \item this model is stratified on ampm
  \item contains a quadratic term on text_length
  \item restricted cubic splines on device_length (5 knots), avg_tag_length (4 knots), and newline_ratio (4 knots)
}
}
