% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifixit_research_functions.R
\name{get_au_terms}
\alias{get_au_terms}
\title{Get most frequent terms in the titles of answered/unanswered questions}
\usage{
get_au_terms(data, variable, stopwords = NULL)
}
\arguments{
\item{data}{The full data set}

\item{variable}{The variable to get the frequent terms from. In the model, question titles were used.
Argument should be input as a string.}

\item{stopwords}{Optional, add stopwords to remove. Argument should be input in the form of a string or character vector.
For the cox regression model, "can", "will", "cant", "wont", "works", "get", "help", "need", "fix", "doesnt", "dont" were removed.}
}
\value{
Returns a data frame of words from the user-specified text variable, along with the frequency each word
occurs in all of the data, as well as in answered and unanswered questions, and a ratio calculated as: frequency
in answered divided by frequency in unanswered. The resulting data frame is used in set up functions for the
contain_answered and contain_unanswered variables.
}
\description{
This function is used in variable setup functions for the cox regression model. It separates the full data
into data frames of answered and unanswered questions. It then uses the get_freq_terms function from this
package to get data frames of the most commonly used words in the user-specified text variable of
answered and unanswered questions (the fitted model used question titles). The resulting data frames are then
joined by word.
}
\examples{
removethese <- c("can", "will", "cant", "wont", "works", "get", "help", "need", "fix", "doesnt", "dont")
get_au_terms(data = x, variable = "title", stopwords = removethese)

}
