% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory_analysis.R
\name{get_avgcv}
\alias{get_avgcv}
\title{Computes the average of performance metrics for training and test sets}
\usage{
get_avgcv(list)
}
\arguments{
\item{list}{list of data frames, output from cv function}
}
\value{
data frame with averages of metrics
}
\description{
Takes output from cv function in the form of a list, and computes averages across training and test sets
}
