% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison.R
\name{get_rsq_adj}
\alias{get_rsq_adj}
\title{Compute r-square adjusted}
\usage{
get_rsq_adj(cox, k)
}
\arguments{
\item{cox}{Cox regression object to find the r-square adjusted value for}

\item{k}{Number of independent predictors in the model}
}
\value{
r-square adjusted value
}
\description{
Computes the r-square adjusted value for cox regression objects
}
