% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R, R/model_comparison.R
\name{setup}
\alias{setup}
\alias{setup}
\title{Set up iFixit Answers data}
\usage{
setup()

setup()
}
\value{
data frame

data frame
}
\description{
Subsets iFixit Answers data to questions in English. Sets up time_until_answer (hrs) variable.

Subsets iFixit Answers data to questions in English. Sets up time_until_answer (hrs) variable.
}
