% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequent_terms .R
\name{clean_corpus}
\alias{clean_corpus}
\title{Cleans corpus}
\usage{
clean_corpus(corpus, stopwrds = NULL)
}
\arguments{
\item{corpus}{corpus to be cleaned}

\item{stopwrds}{optional, adds stopwords to remove}
}
\value{
corpus
}
\description{
Cleans corpus
}
