% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequent_terms.R
\name{clean_corpus}
\alias{clean_corpus}
\title{Clean corpus}
\usage{
clean_corpus(corpus, stopwords = NULL)
}
\arguments{
\item{corpus}{corpus to be cleaned. To create the corpus use: source <- VectorSource(vec), corpus <- VCorpus(source)}

\item{stopwords}{optional, adds stopwords to remove. If not specified it will only remove English stopwords from the tm package.}
}
\value{
corpus
}
\description{
Removes whitespace, removes punctuation, transforms characters to lowercase, and removes stopwords (with the option to add additional stopwords). Used in get_freq_terms.
}
