% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research_functions.R
\name{compare_nested}
\alias{compare_nested}
\title{Compare nested models}
\usage{
compare_nested(full_model, reduced_model)
}
\arguments{
\item{full_model}{Cox regression object (from the function, coxph) for the full model.}

\item{reduced_model}{Cox Regression object for the reduced model.}
}
\value{
data frame with the statistic, degrees of freedom and p-value.
}
\description{
Performs partial likelihood ratio test for cox regression objects using evaluated
log partial likelihoods from the full and nested models to determine if the inclusion
of additional predictors in the model is necessary. This function only works for coxph fits.
}
