% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_failure_functions.R
\name{fit_model}
\alias{fit_model}
\title{Fit cox regression model}
\usage{
fit_model(data, summary = FALSE)
}
\arguments{
\item{data}{The data to use in fitting the model.}

\item{summary}{If true, this function will print a summary of the model (e.g. statistics, coefficients).
Default is set to false.}
}
\value{
Returns the cox regression model
}
\description{
This function sets up variables in the input data required to fit the model using
the variable_setup function from this package, and then fits and returns the model.
}
\details{
This model fits a cox regression model with 19 independent predictors. See the documentation on the
variable_setup function for details on what each predictor represents. This model is stratified on ampm, contains
a quadratic term on text_length, and restricted cubic splines on device_length, avg_tag_length, and newline_ratio with
5, 4, 4 knots, respectively.
}
