% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequent_terms .R
\name{get_freq_terms}
\alias{get_freq_terms}
\title{Returns a data frame of terms and their frequencies}
\usage{
get_freq_terms(vec, n = NULL, stopwrds = NULL)
}
\arguments{
\item{vec}{character vector to be turned into corpus}

\item{n}{optional, specifies the number of frequent terms to return}

\item{stopwrds}{optional, adds stopwards to remove}
}
\value{
data frame
}
\description{
Returns a data frame of terms and their frequencies
}
