% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequent_terms.R
\name{get_freq_terms}
\alias{get_freq_terms}
\title{Term frequency}
\usage{
get_freq_terms(vec, n = NULL, stopwords = NULL)
}
\arguments{
\item{vec}{character vector to be turned into corpus}

\item{n}{optional, specifies the number of frequent terms to return. If not specified it will return the entire data frame.}

\item{stopwords}{optional, adds stopwards to remove. If not specified it will only remove English stopwords from the tm package.}
}
\value{
data frame
}
\description{
Turns the input character vector into a DocumentTermMatrix. Sums the frequencies of each word. Returns a data frame with terms and frequencies arranged from most to least frequent.
}
