% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_to_predict.R
\name{predict_failure}
\alias{predict_failure}
\title{Predict failure probabilities for questions on Answers}
\usage{
predict_failure(model, newdata = NULL, times = c(0.5, 1, 3, 5, 10, 24))
}
\arguments{
\item{model}{The cox regression model to use for predictions (output from the fit_model function). This function
only works with cph fits, not coxph fits.}

\item{newdata}{Optional, new data from which to get predictions for. If this is omitted, this function
will output predictions for each row/question in the data set the model was fit on.}

\item{times}{vector of times at which to predict on. If omitted, this function will return predictions at
0.5, 1, 3, 4, 10, 24 hours.}
}
\value{
Returns a data frame of predicted failure probabilities. The columns are the times predicted on, the
rows correspond to each question in the data.
}
\description{
This function uses a fitted cox proportional hazards model to predict failure probabilities with the
survest function from the rms package. Failure is defined as 1 - survival probability, and indicates the
probability that an event does happen before a certain time. In this case, the failure probability at time t
for a question is the probability that a question receives an answer before time t.
}
