% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifixit_research_functions.R
\name{setup}
\alias{setup}
\title{Set up iFixit Answers data}
\usage{
setup()
}
\value{
data frame
}
\description{
Subsets iFixit Answers data to only questions in English. Sets up time_until_answer (hrs) variable. Creates other predictor variables used in model building.
}
