% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifixit_research_functions.R
\name{variable_setup}
\alias{variable_setup}
\title{Variable setup for final model}
\usage{
variable_setup(data)
}
\arguments{
\item{data}{data set to use}
}
\value{
data frame
}
\description{
Subsets data to all questions in English. Sets up all variables included in the final model.
}
