% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_failure_functions.R
\name{variable_setup}
\alias{variable_setup}
\title{Set up variables in iFixit Answers data for the cox regression model}
\usage{
variable_setup(data)
}
\arguments{
\item{data}{Answers data frame (also include what kind of variables should be included in the data)}
}
\value{
data frame
}
\description{
Sets up all variables needed. This function is used in fit_model.
}
\details{
explain the variables it sets up
}
